backend/foodgram/api/mixin.py

from rest_framework import mixins, viewsets


class MyCreateDestroyClass(mixins.DestroyModelMixin,
                            viewsets.GenericViewSet):
    "Кастомный миксин класс"
    pass